+++ R/op-rng_bit_generator.R
@@ -31,7 +31,7 @@ infer_types_rng_bit_generator <- function(
   shape
 ) {
   assert_vt_is_tensor(initial_state)
-  assert_vt_has_ttype(initial_state, dtype = UIntegerType(64L), naxes = 1L)
+  assert_vt_has_ttype(initial_state, dtype = as_dtype("ui64"), naxes = 1L)
 
   if (!test_choice(rng_algorithm, c("DEFAULT", "THREE_FRY", "PHILOX"))) {
     cli_abort(c(
@@ -61,9 +61,9 @@ infer_types_rng_bit_generator <- function(
   }
 
   out_dtype <- as_dtype(dtype)
-  assert_one_of(
+  assert_dtype_one_of(
     out_dtype,
-    c("IntegerType", "UIntegerType", "FloatType")
+    c("int", "uint", "float")
   )
   out_shape <- as.integer(shape)
 
@@ -71,7 +71,7 @@ infer_types_rng_bit_generator <- function(
   ValueTypes(list(
     ValueType(
       TensorType(
-        dtype = UIntegerType(64L),
+        dtype = as_dtype("ui64"),
         shape = Shape(init_shape)
       )
     ),